Refactor custom address softing
