+++ HousingSearchApi/V1/Helper/AddressComparer.cs
@@ -1,5 +1,4 @@
 using Hackney.Shared.HousingSearch.Domain.Asset;
-using System;
 using System.Collections.Generic;
 using System.Linq;
